# backend/openapi.json
{"openapi":"3.0.0","info":{"title":"Real-Time Document Editor API","version":"1.0.0","description":"API for collaborative document editing"},"servers":[{"url":"http://localhost:3000","description":"Local server"},{"url":"https://my-live-demo.vercel.app","description":"Production server"}],"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"User":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","USER"]},"createdAt":{"type":"string","format":"date-time"}}},"Document":{"type":"object","properties":{"id":{"type":"integer"},"title":{"type":"string"},"content":{"type":"string"},"version":{"type":"integer"},"owner_id":{"type":"integer"},"createdAt":{"type":"string","format":"date-time"}}},"Token":{"type":"object","properties":{"token":{"type":"string"}}},"Login":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}}},"DocumentCreate":{"type":"object","required":["title"],"properties":{"title":{"type":"string"},"content":{"type":"string"}}},"DocumentUpdate":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"version":{"type":"integer"}}},"ShareRequest":{"type":"object","required":["email","permission"],"properties":{"email":{"type":"string","format":"email","example":"user@example.com"},"permission":{"type":"string","enum":["view","edit"],"example":"edit"}}}}},"security":[{"BearerAuth":[]}],"paths":{"/auth/register":{"post":{"summary":"Register a new user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email","example":"user@example.com"},"password":{"type":"string","format":"password","example":"password123!"}}}}}},"responses":{"201":{"description":"User created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"400":{"description":"Invalid input"}}}},"/auth/login":{"post":{"summary":"Login existing user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}}},"responses":{"200":{"description":"Successful login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"401":{"description":"Invalid credentials"}}}},"/documents/create-document":{"post":{"summary":"Create new document","tags":["Documents"],"security":[{"BearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentCreate"}}}},"responses":{"201":{"description":"Document created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Document"}}}}}}},"/documents/documents":{"get":{"summary":"Get all accessible documents","tags":["Documents"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"List of documents","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Document"}}}}}}}},"/documents/{id}":{"get":{"summary":"Get document by ID","tags":["Documents"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true}],"responses":{"200":{"description":"Document details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Document"}}}},"403":{"description":"Unauthorized access"}}},"delete":{"summary":"Delete a document","tags":["Documents"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"The ID of the document to delete"}],"responses":{"204":{"description":"Document deleted successfully"},"403":{"description":"Not document owner"},"404":{"description":"Document not found"}}}},"/documents/edit/{id}":{"put":{"summary":"Update document content","tags":["Documents"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentUpdate"}}}},"responses":{"200":{"description":"Updated document","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Document"}}}},"409":{"description":"Version conflict"}}}},"/documents/{id}/share":{"post":{"summary":"Share document with another user via email","tags":["Documents"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"ID of the document to share"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShareRequest"}}}},"responses":{"201":{"description":"Document shared successfully"},"400":{"description":"Invalid input (e.g., invalid permission)"},"403":{"description":"User is not the document owner"},"404":{"description":"Target user not found"}}}},"/users/users":{"get":{"summary":"Get all users (admin only)","tags":["Users"],"security":[{"BearerAuth":[]}],"responses":{"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"403":{"description":"Requires admin role"}}}},"/users/{id}":{"delete":{"summary":"Delete user (admin only)","tags":["Users"],"security":[{"BearerAuth":[]}],"parameters":[{"in":"path","name":"id","schema":{"type":"integer"},"required":true}],"responses":{"204":{"description":"User deleted"},"403":{"description":"Cannot delete self"},"404":{"description":"User not found"}}}}},"tags":[{"name":"Authentication","description":"User registration and login"},{"name":"Documents","description":"Document management"},{"name":"Users","description":"User management (admin only)"}]}